New String.Manipulation.cs file, move Trim* functions
