+++ src/System.Private.CoreLib/src/System.Private.CoreLib.csproj
@@ -439,6 +439,7 @@
     <Compile Include="System\Security\VerificationException.cs" />
     <Compile Include="System\Single.cs" />
     <Compile Include="System\String.cs" />
+    <Compile Include="System\String.Manipulation.cs" />
     <Compile Include="System\StringComparer.cs" />
     <Compile Include="System\Array.cs" />
     <Compile Include="System\MDArray.cs" Condition="'$(IsProjectNLibrary)' == 'true'" />
